+++ app/controllers/api/users/tasks_controller.rb
@@ -10,7 +10,9 @@ class Api::Users::TasksController < ApiController
   end
 
   def create
-    @task = current_user.tasks.create!(create_task_params)
+    @task = current_user.tasks.new(create_task_params)
+    @task.sync_state_with_project
+    @task.save!
     render status: :created
   end
 
@@ -18,13 +20,9 @@ private
 
   def create_task_params
     parameters = fetch_resource_params(:task, [:label], [:planned_at, :project_id])
-    project = Project.find(parameters[:project_id]) if parameters.has_key?(:project_id)
     if parameters.has_key?(:planned_at)
       parameters[:state] = 'planned'
       parameters[:planned_at] = parameters[:planned_at].to_datetime
-      parameters[:started_at] = DateTime.now
-    elsif project.nil? || project.started?
-      parameters[:state] = 'started'
       parameters[:started_at] = DateTime.now
     else
       parameters[:state] = 'newed'