+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### API changes
 
 - Remove project attributes on task relationships
+- Task's state is synced with its related project on `PATCH /api/tasks/:id`
+  (please refer to tasks API documentation to know more about it)
 
 ### Migration notes